feat tutor schedule v3 change integer

+++ src/main/java/com/learning/api/entity/TutorSchedule.java
@@ -19,8 +19,8 @@ public class TutorSchedule {
     private Long tutorId;
 
     @Column(nullable = false)
-    private Byte weekday;
+    private Integer weekday;
 
     @Column(nullable = false)
-    private Byte hour;
+    private Integer hour;
 }
